backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
alembic==1.14.0
yfinance==1.1.0
python-dotenv==1.0.1
boto3==1.35.94
python-jose[cryptography]==3.3.0
requests==2.32.3
